Move and rename envisalink base entity to separate module (#126054)
