+++ homeassistant/components/envisalink/binary_sensor.py
@@ -13,14 +13,8 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 from homeassistant.util import dt as dt_util
 
-from . import (
-    CONF_ZONENAME,
-    CONF_ZONETYPE,
-    DATA_EVL,
-    SIGNAL_ZONE_UPDATE,
-    ZONE_SCHEMA,
-    EnvisalinkDevice,
-)
+from . import CONF_ZONENAME, CONF_ZONETYPE, DATA_EVL, SIGNAL_ZONE_UPDATE, ZONE_SCHEMA
+from .entity import EnvisalinkEntity
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -52,7 +46,7 @@ async def async_setup_platform(
     async_add_entities(entities)
 
 
-class EnvisalinkBinarySensor(EnvisalinkDevice, BinarySensorEntity):
+class EnvisalinkBinarySensor(EnvisalinkEntity, BinarySensorEntity):
     """Representation of an Envisalink binary sensor."""
 
     def __init__(self, hass, zone_number, zone_name, zone_type, info, controller):
