+++ homeassistant/components/envisalink/sensor.py
@@ -16,8 +16,8 @@ from . import (
     PARTITION_SCHEMA,
     SIGNAL_KEYPAD_UPDATE,
     SIGNAL_PARTITION_UPDATE,
-    EnvisalinkDevice,
 )
+from .entity import EnvisalinkEntity
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -49,7 +49,7 @@ async def async_setup_platform(
     async_add_entities(entities)
 
 
-class EnvisalinkSensor(EnvisalinkDevice, SensorEntity):
+class EnvisalinkSensor(EnvisalinkEntity, SensorEntity):
     """Representation of an Envisalink keypad."""
 
     def __init__(self, hass, partition_name, partition_number, info, controller):
